Move kubeconfig helpers to separate package
